requirements.txt
biopython==1.83
sentence-transformers==2.2.2
transformers==4.30.2
huggingface_hub==0.14.1
scikit-learn==1.3.0
numpy==1.24.4
scipy==1.11.4
python-dotenv==1.0.1
matplotlib==3.8.2
umap-learn==0.5.5
keybert==0.7.0
wordcloud==1.9.3
